Refactor DTOs and AutoMapper profiles for clarity
This commit reorganizes the codebase by standardizing the structure of DTO classes under the new namespace `AVMAPP.Data.APi.Models.Dtos`. It updates multiple controllers to reference the new DTOs and modifies `using` directives accordingly.

Several DTO classes, including `CartItemDto`, `CategoryDto`, and others, have been redefined with their properties. Corresponding AutoMapper profiles have been created or updated to facilitate entity-DTO mapping.

Additionally, the `Program.cs` file has been modified to include the new AutoMapper profiles, and the project file for `AVMAPP.Models.DTo` has been corrected to ensure proper references. The solution file has also been updated to fix a typographical error in the project name.

Finally, new package references for AutoMapper have been added to the `AVMAPP.Services` project, enhancing the maintainability and clarity of the codebase.
