+++ AVMAPP.Data.APi/Controllers/CategoryController.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using AVMAPP.Data.APi.Models;
+using AVMAPP.Data.APi.Models.Dtos;
 using AVMAPP.Data.Entities;
 using AVMAPP.Data.Infrastructure;
 using Microsoft.AspNetCore.Http;
