+++ AVMAPP.Data.APi/Controllers/ProductCommentController.cs
@@ -1,10 +1,11 @@
 ﻿using AutoMapper;
-using AVMAPP.Data.APi.Models;
 using AVMAPP.Data.Entities;
 using AVMAPP.Data.Infrastructure;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
+using AVMAPP.Models.DTo;
+using AVMAPP.Data.APi.Models;
 
 namespace AVMAPP.Data.APi.Controllers
 {
