--- AVMAPP.Data.APi/Models/ProductImageDto.cs
@@ -1,30 +1,0 @@
-﻿using AutoMapper;
-using AVMAPP.Data.Entities;
-
-namespace AVMAPP.Data.APi.Models
-{
-    public class ProductImageDto
-    {
-        public int ProductId { get; set; }
-        public ProductEntity? Product { get; set; }
-        public string Url { get; set; } = null!;
-        public DateTime CreatedAt { get; set; }
-        public DateTime UpdatedAt { get; set; }
-        public bool IsActive { get; set; }
-        public bool IsDeleted { get; set; }
-        public int Id { get; set; }
-    }
-    public class  ProductImageProfile:Profile
-    {
-        public ProductImageProfile()
-        {
-            CreateMap<ProductImageEntity, ProductImageDto>()
-                .ReverseMap()
-                .ForMember(dest => dest.CreatedAt, opt => opt.Ignore())
-                .ForMember(dest => dest.UpdatedAt, opt => opt.Ignore())
-                .ForMember(dest => dest.IsDeleted, opt => opt.Ignore())
-                .ForMember(dest => dest.Product, opt => opt.Ignore());
-        }
-
-    }
-}
