+++ AVMAPP.Moldels.DTo/AVMAPP.Models.DTo.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\AVMAPP.Data\AVMAPP.Data.csproj" />
+  </ItemGroup>
+
 </Project>
