--- AVMAPP.Moldels.DTo/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace AVMAPP.Moldels.DTo
-{
-    public class Class1
-    {
-
-    }
-}
