+++ AVMAPP.Moldels.DTo/Dtos/CategoryDto.cs
@@ -1,0 +1,11 @@
+﻿
+namespace AVMAPP.Data.APi.Models.Dtos
+{
+    public class CategoryDto
+    {
+        public string Name { get; set; }
+        public string Color { get; set; } = "#FFFFFF"; // Varsayılan beyaz renk
+        public string Icon { get; set; } = "icon-avg";     
+
+    }      
+}
