+++ AVMAPP.Moldels.DTo/Dtos/ProductDto.cs
@@ -1,5 +1,4 @@
-﻿using AutoMapper;
-using AVMAPP.Data.Entities;
+﻿using AVMAPP.Data.Entities;
 
 namespace AVMAPP.Data.APi.Models
 {
@@ -19,10 +18,5 @@ namespace AVMAPP.Data.APi.Models
         public bool IsDeleted { get; set; }
         public int Id { get; set; }
     }
-    public class ProductProfile : Profile
-    {
-        public ProductProfile() {
-            CreateMap<ProductEntity, ProductDto>().ReverseMap();
-        }
-    }
+   
 }
