+++ AVMAPP.Services/AVMAPP.Services.csproj
@@ -7,6 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="AutoMapper" Version="15.0.1" />
+    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.18" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="8.0.0" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.2" />
@@ -14,6 +16,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\AVMAPP.Data\AVMAPP.Data.csproj" />
+    <ProjectReference Include="..\AVMAPP.Moldels.DTo\AVMAPP.Models.DTo.csproj" />
   </ItemGroup>
 
 </Project>
