+++ AVMAPP.Services/ServiceExtencions.cs
@@ -1,13 +1,14 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using AVMAPP.Data.Infrastructure;
+using AVMAPP.Data.Infrastructure.AVMDbContext;
+using AVMAPP.Services.Profiles;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
-using AVMAPP.Data.Infrastructure.AVMDbContext;
-using AVMAPP.Data.Infrastructure;
 
 namespace AVMAPP.Services
 {
@@ -21,6 +22,16 @@ namespace AVMAPP.Services
                 options.UseSqlServer(configuration.GetConnectionString("DefaultConnection"));
             });
             services.AddScoped(typeof(IGenericRepository<>), typeof(GenericRepository<>));
+            // Replace 'AllProfile' with the correct AutoMapper profile class name.
+            // If you have a profile class (e.g., 'MappingProfile'), use its type instead.
+            // Example: services.AddAutoMapper(typeof(MappingProfile).Assembly);
+
+            // If you do not have any profile class, you need to create one in AVMAPP.Services.Profiles namespace.
+            // For now, replace 'AllProfile' with the correct profile class name or ask for clarification if you are unsure.
+
+            services.AddAutoMapper(typeof(OrderItemProfile).Assembly);
+      
         }
+        
     }
 }
