+++ AVMAPP.sln
@@ -1,11 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.14.36310.24 d17.14
+VisualStudioVersion = 17.14.36310.24
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AVMAPP.ETicaret.MVC", "AVMAPP.ETicaret.MVC\AVMAPP.ETicaret.MVC.csproj", "{5F052CC6-C388-45C1-90C1-9AB7FA410379}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AVMAPP.Moldels.DTo", "AVMAPP.Moldels.DTo\AVMAPP.Moldels.DTo.csproj", "{3604744A-9E9B-4D3A-914D-918FD9258FC2}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AVMAPP.Models.DTo", "AVMAPP.Moldels.DTo\AVMAPP.Models.DTo.csproj", "{3604744A-9E9B-4D3A-914D-918FD9258FC2}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AVMAPP.Admin.MVC", "AVMAPP.Admin.MVC\AVMAPP.Admin.MVC.csproj", "{D432ECCF-0FAC-472D-A1E7-DA3A5968B6CB}"
 EndProject
